+++ Features/Popup/PopupView.swift
@@ -596,6 +596,7 @@ struct PopupView: View {
         <script>
             window.collapseMode = "\(userConfig.collapseMode.rawValue)";
             window.expandFirstDictionary = \(userConfig.expandFirstDictionary);
+            window.twoColumnLayout = \(userConfig.twoColumnLayout);
             window.collapsedDictionaries = \(collapsedDictionaries);
             window.compactGlossaries = \(userConfig.compactGlossaries);
             window.showExpressionTags = \(userConfig.showExpressionTags);